Add AlarmData getter/setter, Using getter/setter, AlarmData instance val fix
