+++ app/src/main/java/com/ourincheon/wazap/WazapService.java
@@ -16,8 +16,6 @@ import com.ourincheon.wazap.Retrofit.reqContest;
 import retrofit2.Call;
 import retrofit2.http.Body;
 import retrofit2.http.DELETE;
-import retrofit2.http.Field;
-import retrofit2.http.FormUrlEncoded;
 import retrofit2.http.GET;
 import retrofit2.http.Header;
 import retrofit2.http.POST;
